+++ mml/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved the way `FilterParts::Only` works: it can now take a subpart (especially the `text` subpart). When `FilterParts::Only("text")` is given, all text parts are shown as they are, without MML markup (except HTML parts, they are textified). [#487]
+
 ### Fixed
 
 - Fixed missing `compiler` feature guard on error variants.
@@ -238,3 +242,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Imported code from `pimalaya-email-tpl`.
+
+[#487]: https://github.com/pimalaya/himalaya/issues/487